==> 6 model for stage2/01_pre_add_stage1test_to_train.ipynb <==
dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

import numpy as np 
import pandas as pd 

stage1_test = pd.read_csv('../input/test_variants')
stage1_solution = pd.read_csv('../input/stage1_solution_filtered.csv')
stage1_merge = stage1_solution.merge(stage1_test, how = 'left', on = 'ID')

stage1_merge['Class'] = stage1_merge.apply(lambda x:int(x['class1']*1+x['class2']*2+x['class3']*3+x['class4']*4+x['class5']*5\
                        +x['class6']*6+x['class7']*7+x['class8']*8+x['class9']*9), axis=1)

stage1_merge.drop(stage1_merge.columns[[1,2,3,4,5,6,7,8,9]], axis=1,inplace=True) 

stage1_merge.head()

test_txt = pd.read_csv(dir+'/input/test_text', sep="\|\|", \
                        engine='python', header=None, skiprows=1, names=["ID","Text"])

stage1_merge = stage1_merge.merge(test_txt, how = 'left', on = 'ID')

stage1_merge.head()

stage1_merge['ID'] += 10000

stage1_merge.tail()

train_txt = pd.read_csv(dir+'/input/training_text', sep="\|\|", \
                        engine='python', header=None, skiprows=1, names=["ID","Text"])

train_txt.tail()

train_genevar = pd.read_csv(dir+'/input/training_variants')

train_genevar.tail()

add_train_txt = pd.DataFrame(stage1_merge, columns=['ID', 'Text'])
add_train_genevar = pd.DataFrame(stage1_merge, columns=['ID', 'Gene','Variation','Class'])

train_txt_joint = pd.concat([train_txt,add_train_txt], ignore_index=True)


train_txt_joint.tail()

train_genevar_joint = pd.concat([train_genevar,add_train_genevar], ignore_index=True)

train_genevar_joint.tail()

train_genevar_joint.to_csv(dir+'/4 input for stage2/training_variants' + '.csv', index=False)

train_txt_joint.to_csv(dir+'/4 input for stage2/training_txt' + '.csv', index=False)
