==> 6 model for stage2/02_pre_split_test_for_stage2.ipynb <==
dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

import numpy as np 
import pandas as pd 

pre_submission = pd.read_csv('../4 input for stage2/pre_submission.csv')

pre_submission.head()

all_submission = pd.read_csv('../input/stage2_sample_submission.csv')

all_submission.head()

pre_submission = pre_submission.set_index('ID')

all_submission = all_submission.set_index('ID')

other_submission = all_submission.reindex(all_submission.index.difference(pre_submission.index))

other_submission = other_submission.reset_index()

other_submission.head()

stage2_test_genevar = pd.read_csv('../input/stage2_test_variants.csv')
stage2_test_txt = pd.read_csv('../input/stage2_test_text.csv', sep="\|\|", \
                        engine='python', header=None, skiprows=1, names=["ID","Text"])

stage2_test_genevar.head()

stage2_test_txt.head()

df_merge = other_submission.merge(stage2_test_genevar, how = 'left', on = 'ID')

df_merge = df_merge.merge(stage2_test_txt, how = 'left', on = 'ID')

df_merge.head()

test_txt = pd.DataFrame(df_merge, columns=['ID', 'Text'])
test_genevar = pd.DataFrame(df_merge, columns=['ID', 'Gene','Variation'])

test_txt.to_csv(dir+'/4 input for stage2/test_txt' + '.csv', index=False)
test_genevar.to_csv(dir+'/4 input for stage2/test_variants' + '.csv', index=False)

other_submission.to_csv(dir+'/4 input for stage2/other_submission' + '.csv', index=False)
