==> 6 model for stage2/05_feature_egineering_for_variation-Stage2.ipynb <==
import numpy as np
import pandas as pd

dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

train = pd.read_csv(dir+'/4 input for stage2/training_variants.csv')
test = pd.read_csv(dir+'/4 input for stage2/test_variants.csv')
test['Class'] = 0

train.head()

test.head()

print("Train shape".ljust(15), train.shape)
print("Test shape".ljust(15), test.shape)

genes_train = set(train["Gene"])
print("%i train unique Genes" %(len(genes_train)))

genes_test = set(test["Gene"])
print("%i test unique Genes" %(len(genes_test)))

variations_train = set(train["Variation"])
print("%i train unique Variations" %(len(variations_train)))

variations_test = set(test["Variation"])
print("%i test unique Variations" %(len(variations_test)))

genes_shared = genes_train&genes_test
print("%i shared genes in train and test" %(len(genes_shared)))

variations_shared = variations_train&variations_test
print("%i shared variations in train and test" %(len(variations_shared)))

print(variations_shared)

from collections import Counter
top_gene_train = Counter(train["Gene"]).most_common(10)
print(top_gene_train)

top_gene_test = Counter(test["Gene"]).most_common(10)
print(top_gene_test)

top_variation_train = Counter(train["Variation"]).most_common(10)
print(top_variation_train)

top_variation_test = Counter(test["Variation"]).most_common(10)
print(top_variation_test)

# next generate more var info

df_joint = pd.concat([train,test], ignore_index=True)
print("train+test rows:",df_joint.shape[0])


df_joint.tail()

import re
rule_basic = re.compile(r'^([a-z\*])(\d{1,7})([a-z\*])$',)
truncating_list = ['truncating mutations','trunc']
delins_list = ['insertions/deletions','deletion/insertion','delins']
fusion_list = ['fusions','fusion','fus','fs*','fs']
deletion_list = ['deletions','deletion','del']
ins_list = ['insertions','insertion','ins']
amplification_list = ['amplification']
overexpression_list = ['overexpression']
splice_list = ['splice']
duplication_list = ['duplications','duplication','dup']
lookup_list = [truncating_list, delins_list, fusion_list, deletion_list,ins_list, 
               amplification_list, overexpression_list,splice_list, duplication_list ]
operation_list = ['truncating', 'delins','fusion', 'deletion', 'insertion', 'amplification', 
                   'overexpression','splice','duplication']
unused_list = ['mutations','_','-']

def var_to_pattern(variation):
    variation = variation.replace('null','*')
    variation = variation.lower()
    result = {'first_letter':'none','location_number':0,'last_letter':'none','operation':'other','other_text':'none'}
    words = variation.split(' ')
    i = len(words)
    m = re.match(rule_basic,words[0])
    if i == 1 and m:
        #substitution case
        result['operation'] = 'substitution'
        if m.group(1) != '*':
            result['first_letter'] = m.group(1)
        result['location_number'] = m.group(2)
        if m.group(3) != '*':
            result['last_letter'] = m.group(3)
    else:
        flag = False
        temp_txt = variation
        for index,item in enumerate(lookup_list):
            for x in item:
                if x in temp_txt:
                    flag = True
                    result['operation'] = operation_list[index]
                    temp_txt = variation.replace(x,' ')
                    break
            if flag:break
        #for rest text
        if not temp_txt.isspace():
            for x in unused_list:
                temp_txt = temp_txt.replace(x,' ')
            if not temp_txt.isspace():
                temp_txt = temp_txt.strip()
                temp_words = temp_txt.split(' ')
                if len(temp_words) != 1 :
                    if len(temp_words[-1]) < 3:
                        del temp_words[-1]
                temp_txt = ' '.join(temp_words)
                result['other_text'] = temp_txt
        
    return result

pattern = df_joint["Variation"].apply(var_to_pattern)

pattern_df = pd.DataFrame(list(pattern))

done_df = pd.concat([df_joint,pattern_df],axis=1)

done_df.to_csv(dir+'/5 middle for stage2/var_done.csv', index=False)

done_df.tail()
