==> 6 model for stage2/07_feature_egineering_for_one-hot_var_and_length-Stage2.ipynb <==
import numpy as np
import pandas as pd

dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

train_cleantxt = pd.read_csv(dir+'/5 middle for stage2/train_cleantxt.txt')
test_cleantxt = pd.read_csv(dir+'/5 middle for stage2/test_cleantxt.txt')

train_len = len(train_cleantxt)
print(train_len)

df_all = pd.concat((train_cleantxt, test_cleantxt), axis=0, ignore_index=True)

df_all.tail()

variation_done_all = pd.read_csv(dir+'/5 middle for stage2/var_done.csv', )

variation_done_all['CleanText'] = df_all['CleanText']

df_pre_all = variation_done_all.drop(['Class'], axis=1)

df_pre_all['Gene'] = df_pre_all['Gene'].map(str.lower)
df_pre_all['Variation'] = df_pre_all['Variation'].map(str.lower)


df_pre_all.head()

df_pre_all.tail()

# label encoder
from sklearn import preprocessing
for c in df_pre_all.columns:
    if df_pre_all[c].dtype == 'object':
        if c == 'Gene':
            lbl = preprocessing.LabelEncoder()
            df_pre_all[c+'_lbl_enc'] = lbl.fit_transform(df_pre_all[c].values)  
            df_pre_all[c+'_len'] = df_pre_all[c].map(lambda x: len(str(x)))
        elif c == 'Variation':
            lbl = preprocessing.LabelEncoder()
            df_pre_all[c+'_lbl_enc'] = lbl.fit_transform(df_pre_all[c].values)  
            df_pre_all[c+'_len'] = df_pre_all[c].map(lambda x: len(str(x)))
            df_pre_all[c+'_words'] = df_pre_all[c].map(lambda x: len(str(x).split(' ')))
        elif c in ['CleanText','other_text']: 
            df_pre_all[c+'_len'] = df_pre_all[c].map(lambda x: len(str(x)))
            df_pre_all[c+'_words'] = df_pre_all[c].map(lambda x: len(str(x).split(' '))) 

def save(filename, contents): 
    fh = open(filename, 'w') 
    fh.write(contents) 
    fh.close() 

#one-hot for var-first and last letter,and operation
df_pre_all = pd.concat([df_pre_all, pd.get_dummies(df_pre_all['first_letter']).\
                        rename(columns=lambda x: 'firstletter_' + str(x))], axis=1)

df_pre_all = pd.concat([df_pre_all, pd.get_dummies(df_pre_all['last_letter']).\
                        rename(columns=lambda x: 'lastletter_' + str(x))], axis=1)

df_pre_all = pd.concat([df_pre_all, pd.get_dummies(df_pre_all['operation']).\
                        rename(columns=lambda x: 'operation_' + str(x))], axis=1)

df_pre_all['Gene_count'] = df_pre_all.apply(lambda x: str(x['CleanText']).count(str(x['Gene'])), axis=1)

df_pre_all['var_count'] = df_pre_all.apply(lambda x: str(x['CleanText']).count(str(x['first_letter'])+\
                                                            str(x['location_number'])+str(x['last_letter'])), axis=1)

df_pre_all['font_var_count'] = df_pre_all.apply(lambda x: str(x['CleanText']).count(str(x['first_letter'])+\
                                                            str(x['location_number'])), axis=1)

df_pre_all['last_var_count'] = df_pre_all.apply(lambda x: str(x['CleanText']).count(str(x['location_number'])+\
                                                            str(x['last_letter'])), axis=1)

# get operation count
truncating_list = ['trunc']
delins_list = ['insertion/deletion','deletion/insertion','delins']
fusion_list = ['fusion']
deletion_list = ['deletion']
ins_list = ['insertion']
amplification_list = ['amplifica']
overexpression_list = ['overexpress']
splice_list = ['splice']
duplication_list = ['duplicat']
substitution_list = ['substitut']

operation_dict = {'truncating':truncating_list, 'delins':delins_list,'fusion':fusion_list, \
                  'deletion':deletion_list, 'insertion':ins_list, 'amplification':amplification_list, \
'overexpression':overexpression_list,'splice':splice_list,'duplication':duplication_list,'substitution':substitution_list}

def get_oper_count(doc, operation):
    count = 0
    if operation != 'other':
        for x in operation_dict[operation]:
            count += doc.count(x)
    return count

df_pre_all['oper_count'] = df_pre_all.apply(lambda x: get_oper_count(str(x['CleanText']),str(x['operation'])), axis=1)

# get othertxt count

def get_othertxt_count(doc,othertxt):
    count = 0
    if othertxt != 'none':
        for x in othertxt.split(' '):
            count += doc.count(x)
    return count

df_pre_all['other_count'] = df_pre_all.apply(lambda x: get_othertxt_count(str(x['CleanText']),str(x['other_text'])), axis=1)

# combine var
def combinevar(first,middle,last):
    if middle == 0:
        return 'none'
    elif last == 'none':
        return str(first)+str(middle)
    elif first == 'none':
        return 'null'+str(middle)+str(last)
    else:
        return str(first)+str(middle)+str(last)
    

df_pre_all['temp_var'] = df_pre_all.apply(lambda x: combinevar(x['first_letter'],x['location_number'],x['last_letter']), axis=1)

#gen_var_lst  in train 
gen_var_lst = sorted(list(df_pre_all[:train_len].Gene.unique()) + list(df_pre_all[:train_len].temp_var.unique()))
print(len(gen_var_lst))

gen_var_lst.remove('none')
print(len(gen_var_lst))

gen_var_str_convert = ' '.join(gen_var_lst)
def save(filename, contents): 
    fh = open(filename, 'w') 
    fh.write(contents) 
    fh.close() 
save(dir+'/5 middle for stage2/gen_var_str.txt', gen_var_str_convert) 

import pickle
pickle.dump(gen_var_lst, open(dir+'/5 middle for stage2/'+'gen_var_lst', 'wb'))

df_done = df_pre_all.drop(['temp_var','ID','Gene','Variation','first_letter','last_letter','operation',\
                          'other_text'],axis=1)
df_done.tail()

import pickle
pickle.dump(df_done, open(dir+'/5 middle for stage2/'+'onehotvar_cleantxt_data', 'wb'))
