==> 6 model for stage2/09_result_ensemble_submission_three.ipynb <==
import numpy as np
import pandas as pd

dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

s1 = pd.read_csv(dir+'/result/stage2/double_valca_lightgbm.csv')

s1 = s1.set_index('ID')

s2 = pd.read_csv(dir+'/result/stage2/double_valca_svm.csv')

s2 = s2.set_index('ID')

s3 = pd.read_csv(dir+'/result/stage2/double_valca_e0.02_depth6_delta5.csv')
s3 = s3.set_index('ID')

s = (s1 + s2 + s3)/3

s = s.reset_index()

s.head()

s.to_csv(dir+'/result/stage2/ensemble_three.csv', index=False)
