==> 6 model for stage2/10_result_merge_submission_final.ipynb <==
import numpy as np
import pandas as pd

dir ='/root/JupyterNotebook/kaggle-Redefining-Cancer-Treatment'

pre_submission = pd.read_csv('../4 input for stage2/pre_submission.csv')

print(len(pre_submission))

cal_submission = pd.read_csv(dir+'/result/stage2/ensemble_three.csv')

print(len(cal_submission))

cal_submission = cal_submission.set_index('ID')
cal_submission = cal_submission.clip(0.05,0.95)
cal_submission = cal_submission.reset_index()

merge_submission = pd.concat((pre_submission, cal_submission), axis=0, ignore_index=True)

merge_submission = merge_submission.sort_values(by='ID')

merge_submission.head()

merge_submission.to_csv(dir+'/result/stage2/final_average.csv', index=False)
